Fix user group deletion events compatibility

+++ administrator/components/com_users/src/Model/GroupModel.php
@@ -11,6 +11,8 @@
 namespace Joomla\Component\Users\Administrator\Model;
 
 use Joomla\CMS\Access\Access;
+use Joomla\CMS\Event\Model\AfterDeleteEvent;
+use Joomla\CMS\Event\Model\BeforeDeleteEvent;
 use Joomla\CMS\Factory;
 use Joomla\CMS\Form\Form;
 use Joomla\CMS\Language\Text;
@@ -239,15 +241,17 @@ class GroupModel extends AdminModel
     public function delete(&$pks)
     {
         // Typecast variable.
-        $pks    = (array) $pks;
-        $user   = $this->getCurrentUser();
-        $groups = Access::getGroupsByUser($user->get('id'));
+        $pks        = (array) $pks;
+        $user       = $this->getCurrentUser();
+        $groups     = Access::getGroupsByUser($user->get('id'));
+        $context    = $this->option . '.' . $this->name;
+        $dispatcher = $this->getDispatcher();
 
         // Get a row instance.
         $table = $this->getTable();
 
         // Load plugins.
-        PluginHelper::importPlugin($this->events_map['delete']);
+        PluginHelper::importPlugin($this->events_map['delete'], null, true, $dispatcher);
 
         // Check if I am a Super Admin
         $iAmSuperAdmin = $user->authorise('core.admin');
@@ -279,7 +283,18 @@ class GroupModel extends AdminModel
 
                 if ($allow) {
                     // Fire the before delete event.
-                    Factory::getApplication()->triggerEvent($this->event_before_delete, [$table->getProperties()]);
+                    $beforeDeleteEvent = new BeforeDeleteEvent($this->event_before_delete, [
+                        'data'    => $table->getProperties(), // @TODO: Remove data argument in Joomla 6, see BeforeDeleteEvent::__constructor()
+                        'context' => $context,
+                        'subject' => $table,
+                    ]);
+                    $result = $dispatcher->dispatch($this->event_before_delete, $beforeDeleteEvent)->getArgument('result', []);
+
+                    if (\in_array(false, $result, true)) {
+                        $this->setError($table->getError());
+
+                        return false;
+                    }
 
                     if (!$table->delete($pk)) {
                         $this->setError($table->getError());
@@ -288,7 +303,13 @@ class GroupModel extends AdminModel
                     }
 
                     // Trigger the after delete event.
-                    Factory::getApplication()->triggerEvent($this->event_after_delete, [$table->getProperties(), true, $this->getError()]);
+                    $dispatcher->dispatch($this->event_after_delete, new AfterDeleteEvent($this->event_after_delete, [
+                        'data'           => $table->getProperties(), // @TODO: Remove data argument in Joomla 6, see AfterDeleteEvent::__constructor()
+                        'deletingResult' => true, // @TODO: Remove data argument in Joomla 6, see AfterDeleteEvent::__constructor()
+                        'errorMessage'   => $this->getError(), // @TODO: Remove data argument in Joomla 6, see AfterDeleteEvent::__constructor()
+                        'context'        => $context,
+                        'subject'        => $table,
+                    ]));
                 } else {
                     // Prune items that you can't change.
                     unset($pks[$i]);
